+++ R/util.r
@@ -48,7 +48,9 @@ standardise_difftime_names <- function(x) {
 }
 
 standardise_period_names <- function(x) {
-  dates <- c("second", "minute", "hour", "day", "week", "month", "year")
+  dates <- c("second", "minute", "hour", "day", "week", "month", "year",
+             ## these ones are used for rounding only
+             "bimonth", "quarter", "halfyear")
   y <- gsub("(.)s$", "\\1", x)
   y <- substr(y, 1, 3)
   res <- dates[pmatch(y, dates)]
@@ -71,6 +73,27 @@ standardise_lt_names <- function(x) {
          call. = FALSE)
   }
   res
+}
+
+## return list(n=nr_untis,  unti="unit_name")
+parse_period_unit <- function(unit) {
+  m <- regexpr(" *(?<n>[0-9.,]+)? *(?<unit>[^ \t\n]+)", unit[[1]], perl = T)
+  if(m > 0){
+    ## should always match
+    nms <- attr(m, "capture.names")
+    nms <- nms[nzchar(nms)]
+    start <- attr(m, "capture.start")
+    end <- start + attr(m, "capture.length") - 1L
+    n <- if(end[[1]] >= start[[1]]){
+           as.integer(str_sub(unit, start[[1]], end[[1]]))
+         } else {
+           1
+         }
+    unit <- str_sub(unit, start[[2]], end[[2]])
+    list(n = n, unit = unit)
+  } else {
+    stop(sprintf("Invalid unit specification '%s'", unit))
+  }
 }
 
 undefined_arithmetic <- function(e1, e2){