+++ tests/testthat/test-parsers.R
@@ -490,8 +490,8 @@ test_that("Quarters are parsed correctly", {
   qs <- c("2016.1", "2016:2", "2016-3", "2016 4", "2016.5")
   out <- ymd(c("2016-01-01 UTC", "2016-04-01 UTC", "2016-07-01 UTC", "2016-10-01 UTC", NA), tz = "UTC")
   expect_equal(parse_date_time2(qs, orders = "Yq"), out)
-  expect_equal(parse_date_time(qs, orders = "Yq"), out)
-  expect_equal(yq(qs, tz = "UTC"), out)
+  expect_equal(parse_date_time(qs, orders = "Yq", quiet = T), out)
+  expect_equal(yq(qs, tz = "UTC", quiet = T), out)
   expect_equal(yq("16.1", "17.3", "2016.1"), ymd(c("2016-01-01", "2017-07-01", "2016-01-01")))
 })